[#116] CompositionalLayout에 필요한 init extension 분리
